--Centralized Cruise Database - Data Integration SOP Documentation Update--
(this version uses version 0.4 of the database)

docs\Centralized Cruise Database - Data Integration SOP.docx was substantially updated and now includes examples from the centralized CTD project of integration with the centralized cruise database.
docs\Centralized Cruise Database - Technical Documentation.docx was updated to fix a link
docs\todo.txt was updated with a note about using roles to grant permissions on the CCD, but this is problematic because roles are not enabled by default when executing PL/SQL code.  This needs more research